07_LLM_Knowledge_Base/kb_app/requirements.txt
streamlit
openai
python-dotenv
langchain # For text splitting and potentially simplified LLM interaction 